requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-marshmallow
psycopg2-binary
marshmallow
marshmallow-sqlalchemy
passlib
python-dotenv
flask-cors
faker
pytest